pyproject.toml
[project]
name = "claude-skill-manager"
version = "0.1.0"
description = "Local-first operating layer for Claude Code skills, plugins, hooks, MCP servers, subagents, and workflow assets"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "sqlmodel>=0.0.22",
    "typer[all]>=0.15.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.2",
    "httpx>=0.28.0",
    "networkx>=3.4",
    "python-dateutil>=2.9.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "types-pyyaml>=6.0.0",
    "types-python-dateutil>=2.9.0",
    "httpx>=0.28.0",
]

[project.scripts]
csm = "apps.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages/core/csm_core", "packages/connectors/csm_connectors", "packages/agents/csm_agents", "apps/api", "apps/cli"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
